Generally complete - see list of TODO's (biggest unknown is ffmpeg not exiting)

+++ transcoder/transcoder/Program.cs
@@ -19,21 +19,27 @@ namespace transcoder
 
                 while (true)
                 {
-                    string fileToTranscodePath = GetFileToTranscode();
-
-                    string transcodedFilePath = TranscodeFile(fileToTranscodePath);
-
-                    if (!String.IsNullOrEmpty(transcodedFilePath))
+                    FileToTranscode fileToTranscode = GetFileToTranscode();
+                    if (fileToTranscode != null)
                     {
-                        bool ok = UploadFileToServer(transcodedFilePath);
+                        string fileToTranscodePath = fileToTranscode.Path;
 
-                        if (ok)
+                        string transcodedFilePath = TranscodeFile(fileToTranscodePath);
+
+                        if (!String.IsNullOrEmpty(transcodedFilePath))
                         {
-                            // delete local files (downloaded file and converted file)
-                            File.Delete(fileToTranscodePath);
-                            File.Delete(transcodedFilePath);
+                            bool ok = UploadFileToServer(fileToTranscode.Id, transcodedFilePath);
+
+                            if (ok)
+                            {
+                                // delete local files (downloaded file and converted file)
+                                File.Delete(fileToTranscodePath);
+                                File.Delete(transcodedFilePath);
+
+                                Trace.WriteLine("Great success!");
+                            }
+                            // delay some amount of time before looking for the next file
                         }
-                        // delay some amount of time before looking for the next file
                     }
                 }
             }
@@ -55,7 +61,7 @@ namespace transcoder
             }
         }
 
-        private static string GetFileToTranscode()
+        private static FileToTranscode GetFileToTranscode()
         {
             HTTPGet httpGet = new HTTPGet();
 
@@ -72,23 +78,50 @@ namespace transcoder
                 XmlNodeList nodes = doc.GetElementsByTagName("FileToTranscode");
                 if (nodes.Count > 0)
                 {
+                    string id = String.Empty;
+                    string relativeUrl = String.Empty;
+
                     XmlElement fileToTranscodeElem = (XmlElement)nodes[0];
 
-                    // XML contains the path of the file relative to root. Use that as the relative url; then use the last part of the relative Url as the file name
-                    string relativeUrl = fileToTranscodeElem.InnerText;
-                    string tmpPath = System.IO.Path.Combine(_tmpFolder, relativeUrl);
-                    string targetPath = System.IO.Path.Combine(_tmpFolder, System.IO.Path.GetFileName(tmpPath));
-
-                    httpGet = new HTTPGet();
-                    httpGet.RequestToFile("http://" + _bsIPAddress + "/" + relativeUrl, targetPath);
-
-                    if (httpGet.StatusCode == 200)
+                    XmlNodeList childNodes = fileToTranscodeElem.ChildNodes;
+                    foreach (XmlNode childNode in childNodes)
                     {
-                        return targetPath;
+                        if (childNode.Name == "id")
+                        {
+                            id = childNode.InnerText;
+                        }
+                        else if (childNode.Name == "path")
+                        {
+                            relativeUrl = childNode.InnerText;
+                        }
+                    }
+
+                    if (id != String.Empty && relativeUrl != String.Empty)
+                    {
+                        // XML contains the path of the file relative to root. Use that as the relative url; then use the last part of the relative Url as the file name
+                        string tmpPath = System.IO.Path.Combine(_tmpFolder, relativeUrl);
+                        string targetPath = System.IO.Path.Combine(_tmpFolder, System.IO.Path.GetFileName(tmpPath));
+
+                        httpGet = new HTTPGet();
+                        httpGet.RequestToFile("http://" + _bsIPAddress + "/" + relativeUrl, targetPath);
+
+                        if (httpGet.StatusCode == 200)
+                        {
+                            return new FileToTranscode
+                            {
+                                Id = id,
+                                Path = targetPath
+                            };
+                        }
+                        else
+                        {
+                            // TODO - check for errors
+                            return null;
+                        }
                     }
                     else
                     {
-                        // TODO - check for errors
+                        // TODO - log an error
                         return null;
                     }
                 }
@@ -115,24 +148,37 @@ namespace transcoder
                 process.StartInfo.FileName = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "ffmpeg.exe");
                 process.StartInfo.Arguments = ffmpegArgs;
                 process.StartInfo.UseShellExecute = false;
-                process.StartInfo.RedirectStandardError = true;
+                //process.StartInfo.RedirectStandardError = true;
+                process.StartInfo.RedirectStandardError = false;
                 process.StartInfo.CreateNoWindow = true;
 
                 //process.Start();
 
+                Trace.WriteLine("Start ffmpeg");
                 if (!process.Start())
                 {
                     Trace.WriteLine("Error starting");
                     return null;
                 }
-                StreamReader reader = process.StandardError;
-                string line;
-                while ((line = reader.ReadLine()) != null)
-                {
-                    Trace.WriteLine(line);
-                }
+
+                //StreamReader reader = process.StandardError;
+                //string line;
+                //while ((line = reader.ReadLine()) != null)
+                //{
+                //    Trace.WriteLine(line);
+                //}
+
                 process.PriorityClass = ProcessPriorityClass.Normal;
-                process.WaitForExit(60000);
+
+                // TODO - does ffmpeg lock up some times - if yes, is this okay?
+                // no, it's not okay. after timeout, ffmpeg is still running, which is a problem
+                // why is ffmpeg locking up?
+                bool processExited = process.WaitForExit(5000);
+                if (!processExited)
+                {
+                    Trace.WriteLine("Kill ffmpeg");
+                    process.Kill();
+                }
             }
             catch (Exception ex)
             {
@@ -149,7 +195,7 @@ namespace transcoder
             return targetPath;
         }
 
-        private static bool UploadFileToServer(string filePath)
+        private static bool UploadFileToServer(string id, string filePath)
         {
             string fileName = System.IO.Path.GetFileName(filePath);
 
@@ -157,6 +203,7 @@ namespace transcoder
 
             nvc.Add("Destination-Filename", String.Concat("content/", fileName));
             nvc.Add("Friendly-Filename", fileName);
+            nvc.Add("DB-Id", id);
 
             try
             {
@@ -177,4 +224,11 @@ namespace transcoder
             return true;
         }
     }
+
+    class FileToTranscode
+    {
+        public string Id { get; set; }
+        public string Path { get; set; }
+    }
+
 }
